requirements.txt
# altair==5.1.2  - Display library, not directly called
# appnope==0.1.3 - prevents your mac os to go on sleep mode
# asttokens==2.4.1 - annotates Python abstract syntax trees
# attrs==23.1.0 - make python classes great again
# blinker==1.7.0 - allow others to subscripte to events
# cachetools==5.3.2 - provides various memoizing collections and decorators,
# certifi==2023.11.17
# charset-normalizer==3.3.2
# click==8.1.7
# comm==0.2.0 - Jupiter notebook stuff
# debugpy==1.8.0 - not used but legit
# decorator==5.1.1 - not used but legit
# exceptiongroup==1.2.0 - exception manager, I don,t see why we need this
# executing==2.0.1 - Not used
# gitdb==4.0.11 - Not used an no description online...
# GitPython==3.1.40 - Not used an not much description same mainteners as gitdb...
# idna==3.4 - Not used
# importlib-metadata==6.8.0 - Not used
# ipykernel==6.26.0 - Not used
# ipython==8.18.0 - Not used and minrk maintainer
# jedi==0.19.1 - Jedi is a static analysis tool for Python used in ide
# Jinja2==3.1.2 - Not used - not verified and no maintainer shown
# jsonschema==4.20.0  - Not used - not verified and no maintainer shown
# jsonschema-specifications==2023.11.1  - Not used - not verified and no maintainer shown
# jupyter_client==8.6.0 - It also provides client and kernel management APIs for working with kernels.
# jupyter_core==5.5.0 - There is no reason to install this package on its own.
# markdown-it-py==3.0.0 - Not used but seems legit
# MarkupSafe==2.1.3 - Not used 
# matplotlib-inline==0.1.6 - Jupiter stuff
# mdurl==0.1.2
# nest-asyncio==1.5.8
# numpy==1.26.4 - Legit, not used directly
# packaging==23.2
pandas==2.2.2
# parso==0.8.3 - Python parser
# pexpect==4.8.0 - Pexpect is a pure Python module for spawning child applications
# Pillow==10.1.0 - python imaging library
# platformdirs==4.0.0 - When writing desktop application, finding the right location to store user data and configuration varies per platform
plotly==5.24.1
# prompt-toolkit==3.0.41
# protobuf==4.25.1
# psutil==5.9.6
psycopg2-binary==2.9.9
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pyarrow==14.0.1
# pydeck==0.8.1b0
# Pygments==2.17.2
# pytest==8.0.0
# python-dateutil==2.8.2
# python-decouple==3.8
python-dotenv==1.0.1
# pytz==2023.3.post1
# pyzmq==25.1.1
# referencing==0.31.0
requests==2.31.0
# rich==13.7.0
# rpds-py==0.13.1
# six==1.16.0
# smmap==5.0.1
SQLAlchemy==2.0.32
# stack-data==0.6.3
streamlit==1.37.1
# tenacity==8.2.3
# toml==0.10.2
# toolz==0.12.0
# tornado==6.3.3
# traitlets==5.13.0
# typing_extensions==4.8.0
# tzdata==2023.3
# tzlocal==5.2
# urllib3==2.1.0
# validators==0.22.0
watchdog==4.0.0
# wcwidth==0.2.12
# zipp==3.17.0
